+++ ta4j-core/src/main/java/org/ta4j/core/analysis/Returns.java
@@ -67,7 +67,7 @@ public class Returns implements Indicator<Num> {
     private final BarSeries barSeries;
 
     /** The return rates. */
-    private List<Num> values;
+    private final List<Num> values;
 
     /** Unit element for efficient arithmetic return computation. */
     private static Num one;