+++ ta4j-core/src/test/java/org/ta4j/core/criteria/LinearTransactionCostCriterionTest.java
@@ -41,11 +41,11 @@ import org.ta4j.core.num.Num;
 
 public class LinearTransactionCostCriterionTest extends AbstractCriterionTest {
 
-    private ExternalCriterionTest xls;
+    private final ExternalCriterionTest xls;
 
     public LinearTransactionCostCriterionTest(Function<Number, Num> numFunction) throws Exception {
-        super((params) -> new LinearTransactionCostCriterion((double) params[0], (double) params[1],
-                (double) params[2]), numFunction);
+        super(params -> new LinearTransactionCostCriterion((double) params[0], (double) params[1], (double) params[2]),
+                numFunction);
         xls = new XLSCriterionTest(this.getClass(), "LTC.xls", 16, 6, numFunction);
     }